+++ lib/Console/Command/ReportCommand.php
@@ -62,15 +62,15 @@ class ReportCommand extends Command
         $this->setHelp(<<<'EOT'
 Generate report from the latest stored suite:
 
-    $ %command.full_name% --uuid=latest --report=aggregate
+    $ %command.full_name% --ref=latest --report=aggregate
 
 Generate report from a tag:
 
-    $ %command.full_name% --uuid=tag:foobar --report=aggregate
+    $ %command.full_name% --ref=foobar-tag --report=aggregate
 
 Generate report from a UUID (as determined from the <info>log</info> command:
 
-    $ %command.full_name% --uuid=133a2605fac74edabf046edeb9c5f7c4dc1a3aac --report=aggregate
+    $ %command.full_name% --ref=133a2605fac74edabf046edeb9c5f7c4dc1a3aac --report=aggregate
 
 Generate from an XML file: