+++ lib/Storage/UuidResolver/TagResolver.php
@@ -14,15 +14,12 @@ namespace PhpBench\Storage\UuidResolver;
 
 use PhpBench\Model\Tag;
 use PhpBench\Storage\Exception\InvalidTagException;
-use PhpBench\Storage\Exception\TagNotFoundException;
 use PhpBench\Storage\HistoryEntry;
 use PhpBench\Storage\StorageRegistry;
 use PhpBench\Storage\UuidResolverInterface;
 
 class TagResolver implements UuidResolverInterface
 {
-    const PREFIX = 'tag:';
-
     /**
      * @var StorageRegistry
      */
@@ -32,21 +29,7 @@ class TagResolver implements UuidResolverInterface
     {
         $this->storageRegistry = $storageRegistry;
     }
-
-    public function supports(string $reference): bool
-    {
-        if (0 === strpos($reference, self::PREFIX)) {
-            if (strlen($reference) === 4) {
-                return false;
-            }
-
-            return true;
-        }
-
-        return false;
-    }
-
-    public function resolve(string $reference): string
+    public function resolve(string $reference): ?string
     {
         $history = $this->storageRegistry->getService()->history();
 
@@ -64,15 +47,7 @@ class TagResolver implements UuidResolverInterface
             }
         }
 
-        if ($offset > 0) {
-            throw new TagNotFoundException(sprintf(
-                'Could not find entry %s entries before tag "%s"', $offset, $tag
-            ));
-        }
-
-        throw new TagNotFoundException(sprintf(
-            'Could not find entry for tag "%s"', $tag
-        ));
+        return null;
     }
 
     /**
@@ -80,7 +55,7 @@ class TagResolver implements UuidResolverInterface
      */
     private function tagAndOffset(string $reference): array
     {
-        if (!preg_match(sprintf('{^tag:(%s)?-?([0-9]+)?$}', Tag::REGEX_PATTERN), $reference, $matches)) {
+        if (!preg_match(sprintf('{^(%s)?-?([0-9]+)?$}', Tag::REGEX_PATTERN), $reference, $matches)) {
             throw new InvalidTagException(sprintf(
                 'Could not parse tag "%s"', $reference
             ));