+++ lib/Storage/UuidResolverInterface.php
@@ -14,7 +14,5 @@ namespace PhpBench\Storage;
 
 interface UuidResolverInterface
 {
-    public function supports(string $reference): bool;
-
-    public function resolve(string $reference): string;
+    public function resolve(string $reference): ?string;
 }