+++ CHANGELOG.md
@@ -9,6 +9,11 @@ Backward compatiblity breaks:
 - `--uuid` renamed to `--ref` and `tag:` prefix removed #740
 - No warnings - if assertion fails within tolerance zone then it is OK
 
+Improvements:
+
+- Show difference to baseline in progress loggers.
+- Highlight assertion failures.
+
 1.0.0-alpha-4
 -------------